backend/torrent/requirements.txt
python-libtorrent
yaspin
